[WFTC-13] LocalUserTransaction provide methods to set it's availability
This is done because of local transaction invoked in scope of
@Transactional interceptor
(e.g. from Narayana cdi implementation) needs to know if it's fine to
being invoked that context if not exception is thrown.

3.7 Transactional Annotation:
If an attempt is made to call any method of the UserTransaction
interface from within the scope of a bean or method annotated with
@Transactional and a Transactional.TxType other than NOT_SUPPORTED or
NEVER, an IllegalStateException must be thrown.