[WFTC-25] reading transaction store when transction is not found in active list
we need to launch recovery to get the transaction from object store being available
for XATerminator import
that could be necessary e.g. in case when container crashes and is
started again. Then it could be contacted to finish in-doubt
SubordinateAtomicAction and info about it is stored only in the tx log
store but not in memory so far.