[WFTC-70]  removing XA file record only on recover start scan
This is a regression from fix WFTC-64 where file descriptor is removed
just when no xids are returned. But the XAResource.recover
can be called with different flags and for particular flags
(e.g. TMENDRSCAN) no xids are expected to be returned.

The file descriptor can be deleted only when real remote call
is processed which is when TMSTARTRSCAN is passed as the flag.